requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
fonttools==4.37.4
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
